Moved galois tests to galois_test mod
